pyproject.toml
[project]
name = "vibeify-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Chris Rinaldi",email = "crinaldi44@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "sqlmodel (>=0.0.14,<1.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.5.0,<3.0.0)",
    "querymate (>=0.6.9,<0.7.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.12,<1.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.0,<6.0.0)",
    "aioboto3 (>=15.5.0,<16.0.0)"
]

[tool.poetry]
name = "vibeify-api"
version = "0.1.0"
description = ""
authors = ["Chris Rinaldi <crinaldi44@gmail.com>"]
readme = "README.md"
packages = [{include = "vibeify_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.13,<0.15.0)",
    "watchfiles (>=0.24.0,<1.0.0)",
    "coverage (>=7.13.1,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)"
]
